Update client.pl
v3.57 (i forgot to put v3.56 here, only difference iirc there was the addition of --verify-x509-name)

The only only change in 3.57 is we no longer need to account for the "oregon-maine" CN in the oregon/maine server's OpenVPN certificate. The reason it was like that is because the Oregon node had a LOT of IPs, and it was causing problems with some people's DNS implementation. Since about half of the IPs on Oregon appeared to be in Maine to most GeoIP databases, we broke those off into a separate maine.cstorm.is, even though we don't have a server in Maine. We've replaced the Oregon server with one that has a lot less IPs, but better speeds, so there's no reason to keep around that Maine "server", or accommodate for the "oregon-maine" CN since the new oregon server simply uses "cryptostorm oregon server" for the CN, for --verify-x509-name