Update client.pl
v3.57 -> v3.58

No changes to the widget itself, only upgrading the built-in OpenVPN from 2.6.8 to 2.6.10 which addresses CVE-2024-27459, CVE-2024-24974, CVE-2024-27903, and CVE-2024-1305. The first three vulnerabilities are related to the interactive service, which we don't use. The last one applies to the TAP driver, which we do use. OpenSSL also upgraded from 3.1.4 to 3.2.1