Update client.pl
v3.58 -> v3.59

Removed all the killswitch code. It relied on Windows Firewall to allow only connections to CS IPs, but to get those IPs it would resolve all the VPN hosts, which isn't the most efficient way of doing that. Also, with that method, port forwarding gets blocked because all incoming traffic not from a CS IP gets blocked.

Instead of doing it that way, now the killswitch simply does:
--pull-filter ignore "redirect-gateway def1"
--redirect-gateway
From the OpenVPN manual:
"def1 - Use this flag to override the default gateway by using 0.0.0.0/1 and 128.0.0.0/1 rather than 0.0.0.0/0. This has the benefit of overriding but not wiping out the original default gateway."
If we ignore the "--redirect-gateway def1" that the server pushes, and instead use only "--redirect-gateway" then the original gateway isn't restored if the widget or OpenVPN crashes, which is effectively the same as a killswitch, except this way doesn't prevent port forwarding from working.

To restore the default route, the widget will run ipconfig /renew on disconnect and on exit. In case the widget crashes, it also is executed when the widget starts up.