Update lang.txt
Italian translation added, and will be in the next (v3.66) release. If you don't want to wait that long, just replace the lang.txt you have in C:\Program Files (x86)\Cryptostorm Client\user\ with this file (be sure to save the raw .txt file and not the HTML of its GitHub page).